trading_agent/agents/scraping_agent.py
import os
import requests
from dotenv import load_dotenv

load_dotenv()

class ScrapingAgent:
    def __init__(self):
        # Simple scraping agent without external dependencies
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        })

    def scrape_news(self, symbol):
        """Scrape news for a specific symbol using a simple approach."""
        try:
            # Use a simple news API or web search
            # For now, return mock data to avoid dependency issues
            mock_news = [{
                "title": f"Latest market analysis for {symbol}",
                "summary": f"Recent market updates and analysis for {symbol} stock performance.",
                "source": "Market News",
                "timestamp": "2024-01-01T00:00:00Z",
                "sentiment": "neutral"
            }]
            return mock_news
        except Exception as e:
            print(f"❌ News scraping failed: {str(e)}")
            return []

    def scrape_financial_news(self, symbol):
        """Scrape financial news for a specific symbol."""
        return self.scrape_news(symbol)

    def scrape_market_data(self, url):
        """Scrape broader market data from a given URL."""
        try:
            # Simple mock data for market information
            return {
                "market_summary": "Market data temporarily unavailable",
                "timestamp": "2024-01-01T00:00:00Z"
            }
        except Exception as e:
            print(f"❌ Market data scraping failed: {str(e)}")
            return {}